GL - merging

+++ Code/Game/GameLogic/Game_LevelData.cpp
@@ -54,8 +54,10 @@ IObjectData* Game::LevelData::GetObjectAt(int ID) const
 
 void Game::LevelData::GetAllDynamicObjects(Utility::DynamicMemory::DynamicArray<IObjectData*>& mem) const
 {
-	for(int i = 0; i < level->dynamicObjects.Size(); i++)
+
+	for(int i = 0; i < (int)level->dynamicObjects.Size(); i++)
 	{
 		mem[i] = level->dynamicObjects[i];
 	}
+
 }
